resume-parser/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
pdfminer.six==20221105
regex==2023.10.3
python-dateutil==2.8.2
gunicorn==21.2.0 